+++ tests/application/app/views/changes.md
@@ -1,3 +1,6 @@
+#### 14/02/2020
+1. Address bug in request checking if IP6 is local
+
 #### 12/02/2020
 1. New _mobile nav hider
    * Hides the Navigation Bar when Scrolling Up, Shows on down